factored out code from Function Handler
